entity heiurarchy cleaned up and working. additional refactoring of scripting for maintainability.
