+++ data/scripts/XML/TestMode.xml
@@ -4,5 +4,6 @@
 	<Controls filename="data/scripts/XML/KeyMapping.xml"></Controls>
 	<!--<GameWorlds filename="data/scripts/XML/GameWorlds.xml"></GameWorlds>-->
 	<GameWorlds filename="data/scripts/XML/HeirarchyTest.xml"></GameWorlds>
+	<!--<GameWorlds filename="data/scripts/XML/CollisionTest.xml"></GameWorlds>-->
 	<!--<GraphicsSettings filename="data/scripts/XML/TestModeRenderer.xml></GraphicsSettings>-->
 </GameMode>
