+++ src/Engine/Subsystems/EC_SystemType.h
@@ -1,7 +1,6 @@
 #pragma once
 enum class EC_SystemType
 {
-	Hierarchy,
 	Spatial,
 	Transform,
 	Camera,
